fixes for sync
git-svn-id: svn://svn.codewave.de/titanium/MyTunesRSS/trunk@7819 6cd82e91-2702-0410-863d-86aa3370aca5

+++ Resources/mytunesrss.js
@@ -252,6 +252,27 @@ function loadAndDisplayOfflineTracks(parent, album, albumArtist) {
     }
 }
 
+function removeObsoleteTracks(tracks) {
+	var db = Titanium.Database.open("OfflineTracks");
+	var rs = db.execute("SELECT id FROM track");
+	while (rs.isValidRow()) {
+		var trackId = rs.fieldByName("id");
+		var obsolete = true;
+		for (var i = 0; i < tracks.length; i++) {
+			if (trackId === tracks[i].id) {
+				obsolete = false;
+				break;
+			}
+		}
+		if (obsolete) {
+			db.execute("DELETE FROM track WHERE id = ?", trackId);
+			getFileForTrackCache(trackId).deleteFile();
+		}
+		rs.next();
+	}
+	db.close();
+}
+
 function loadAndDisplayTracks(parent, tracksUri) {
     var response = restCall("GET", tracksUri + "?" + TRACK_ATTRIBUTES, {});
     if (response.status / 100 === 2) {
@@ -269,35 +290,43 @@ function loadAndDisplayTracks(parent, tracksUri) {
 var CANCEL_SYNC_AUDIO_TRACKS = false;
 
 function syncTrackAndAdvance(data, index, progressCallback, doneCallback) {
-	cacheTrack(data[index].id, data[index].playbackUri, function() {return !CANCEL_SYNC_AUDIO_TRACKS}, function() {
-		db = Titanium.Database.open("OfflineTracks");
-		db.execute("DELETE FROM track WHERE id = ?", data[index].id);
-		db.execute(
-			"INSERT INTO track (id, name, album, artist, genre, album_artist, image_hash, protected, media_type, time, track_number) VALUES (?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?)",
-			data[index].id,
-			data[index].name,
-			data[index].album,
-			data[index].artist,
-			data[index].genre,
-			data[index].albumArtist,
-			data[index].imageHash,
-			data[index].protected,
-			data[index].mediaType,
-			data[index].time,
-			data[index].trackNumber
-		);
-		db.close();
+	cacheTrack(data[index].id, data[index].playbackUri, function() {return !CANCEL_SYNC_AUDIO_TRACKS}, function(e) {
+		if (e == undefined || e.error == undefined) {
+			if (!getImageCacheFile(data[index].imageHash).exists()) {
+				downloadImage(data[index].imageHash, data[index].imageUri);
+			}
+			if (!getImageCacheFile(data[index].imageHash + "_64").exists()) {
+				downloadImage(data[index].imageHash + "_64", data[index].imageUri + "/size=64");
+			}
+			if (!getImageCacheFile(data[index].imageHash + "_128").exists()) {
+				downloadImage(data[index].imageHash + "_128", data[index].imageUri + "/size=128");
+			}
+			db = Titanium.Database.open("OfflineTracks");
+			db.execute("DELETE FROM track WHERE id = ?", data[index].id);
+			db.execute(
+				"INSERT INTO track (id, name, album, artist, genre, album_artist, image_hash, protected, media_type, time, track_number) VALUES (?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?)",
+				data[index].id,
+				data[index].name,
+				data[index].album,
+				data[index].artist,
+				data[index].genre,
+				data[index].albumArtist,
+				data[index].imageHash,
+				data[index].protected,
+				data[index].mediaType,
+				data[index].time,
+				data[index].trackNumber
+			);
+			db.close();
+		}
 		index++;
 		progressCallback(index * 100 / data.length);
-		if (CANCEL_SYNC_AUDIO_TRACKS || index == data.length) {
+		if (CANCEL_SYNC_AUDIO_TRACKS || index >= data.length) {
 			doneCallback();
 		} else {
 			syncTrackAndAdvance(data, index, progressCallback, doneCallback);
 		}
 	});
-	downloadImage(data[index].imageHash, data[index].imageUri);
-	downloadImage(data[index].imageHash + "_64", data[index].imageUri + "/size=64");
-	downloadImage(data[index].imageHash + "_128", data[index].imageUri + "/size=128");
 }
 
 function loadTracks(tracksUri) {
@@ -435,7 +464,7 @@ function getTcParam() {
 }
 
 function clearImageCache() {
-	var baseDir = getServerBasedCacheDir();;
+	var baseDir = getServerBasedCacheDir();
 	var dir = Titanium.Filesystem.getFile(baseDir, "cache", "images");
 	if (dir.exists()) {
 		dir.deleteDirectory(true);
@@ -443,7 +472,7 @@ function clearImageCache() {
 }
 
 function clearTrackCache() {
-	var baseDir = getServerBasedCacheDir();;
+	var baseDir = getServerBasedCacheDir();
 	var dir = Titanium.Filesystem.getFile(baseDir, "cache", "tracks");
 	if (dir.exists()) {
 		dir.deleteDirectory(true);
@@ -525,7 +554,7 @@ function cacheTrack(id, uri, progressCallback, doneCallback) {
 	var file = getFileForTrackCache(id);
 	if (file.exists()) {
 		if (doneCallback != undefined) {
-			doneCallback();
+			doneCallback({});
 		}
 	} else {
 		var httpClient = Titanium.Network.createHTTPClient();
@@ -534,7 +563,7 @@ function cacheTrack(id, uri, progressCallback, doneCallback) {
 				if (!progressCallback(e.progress * 100)) {
 					httpClient.abort();
 					if (doneCallback != undefined) {
-						doneCallback();
+						doneCallback({aborted:true});
 					}
 				}
 			}
